requirements.txt
fastapi==0.103.2
pandas==2.1.1
pydantic==2.2.0
pydantic_settings==2.0.3
SQLAlchemy==2.0.21
uvicorn==0.23.2
mysqlclient==2.1.1
python-jose==3.3.0
python-multipart==0.0.6